requirements.txt
elasticsearch>=6.0.0,<7.0.0
psycopg2-binary>=2.8.0,<2.9.0